Устранение замечаний и исправление логики приложения.

+++ src/main/java/com/homegarden/store/backend/service/FavoriteServiceImpl.java
@@ -1,12 +1,12 @@
 package com.homegarden.store.backend.service;
 
 import com.homegarden.store.backend.entity.Favorite;
-import com.homegarden.store.backend.exception.ProductNotFoundException;
-import com.homegarden.store.backend.exception.UserNotFoundException;
+import com.homegarden.store.backend.entity.User;
 import com.homegarden.store.backend.repository.FavoriteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 
 @Service
@@ -16,24 +16,23 @@ public class FavoriteServiceImpl implements FavoriteService {
     private final FavoriteRepository favoriteRepository;
     private final ProductService productService;
     private final UserService userService;
+    private final AccessCheckService accessCheckService;
 
     @Override
     public List<Favorite> getAll(Long userId) {
-        if (!userService.existsById(userId)) {
-            throw new UserNotFoundException("User with id " + userId + " doesn't exists");
-        }
+        User user = userService.getById(userId);
+        accessCheckService.checkAccess(user);
 
-        return favoriteRepository.findAllByUser_UserId(userId);
+        return favoriteRepository.findAllByUser(user);
     }
 
     @Override
     public void addToFavorites(Favorite favorite) {
-        if (!userService.existsById(favorite.getUser().getUserId())) {
-            throw new UserNotFoundException("User not found");
-        }
-        if (!productService.existsById(favorite.getProduct().getProductId())) {
-            throw new ProductNotFoundException("Product not found");
-        }
+        User user = userService.getById(favorite.getUser().getUserId());
+        accessCheckService.checkAccess(user);
+
+        productService.getById(favorite.getProduct().getProductId());
+
         if (!favoriteRepository
                 .existsByUser_AndProduct(favorite.getUser(), favorite.getProduct())) {
             favoriteRepository.save(favorite);
@@ -43,6 +42,9 @@ public class FavoriteServiceImpl implements FavoriteService {
     @Override
     @Transactional
     public void removeFromFavorites(Favorite favorite) {
+        User user = userService.getById(favorite.getUser().getUserId());
+        accessCheckService.checkAccess(user);
+
         favoriteRepository.deleteByUserAndProduct(
                 favorite.getUser(),
                 favorite.getProduct());
